requirements.txt
# Minimal requirements for GeoTIFF-only pipeline

# Core numeric + vision
numpy==2.2.6
opencv-python==4.12.0.88

# Geospatial raster + vector
rasterio==1.4.3
shapely==2.1.1

# Config
PyYAML==6.0.2
pyproj==3.7.2          

# Local GCP web UI
Flask==3.0.3

# --- Notes ---
# Add these ONLY if/when you actually need them:
# pandas==2.3.2              # data wrangling / CSVs

# fiona==1.10.1              # shapefile/GPX I/O via GDAL
# geopandas==1.1.1           # geospatial DataFrames on top of pandas
# pyogrio==0.11.1            # fast OGR I/O (alternative to fiona)
